fix(plugs): moved private endpoints to private controller for services module
