--- api/services/plugs/src/services/services.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ServicesController } from './services.controller';
-
-describe('ServicesController', () => {
-  let controller: ServicesController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ServicesController],
-    }).compile();
-
-    controller = module.get<ServicesController>(ServicesController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
