--- api/services/plugs/src/services/services/services.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ServicesService } from './services.service';
-
-describe('ServicesService', () => {
-  let service: ServicesService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ServicesService],
-    }).compile();
-
-    service = module.get<ServicesService>(ServicesService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
